fixed the easy ones
added a trivial notification system to replace GenEvent.stream

remove replacement for genevent

and update to new supervisor child_spec model (except for GenEvent, which doesn't support it. We can look at that when a replacement is agreed)

back to old-school supervision. I left the start_link/1 changes in, as they will work with both