Respect secondary_preferred read preference by falling back to primary
As this read preference is following the exact same code path as the secondary
read preference, this has the side-effect of not being able to make read
operations when only the primary is available.

This change ensures this fallback and simplifies the primary_preferred
implementation to mimic the behaviour, only the other way around.